+++ app/sitemap.js
@@ -8,6 +8,7 @@ export default async function sitemap() {
     "",
     "/articles",
     "/projects",
+    "/roadmap",
     "/about",
     "/privacy",
     ...articles.map(({ slug }) => `/articles/${slug}`),